+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2.6.0
+=====
+
+* (bug) Fix invalid type definition.
+* (feature) Transform data of embedded blocks in RTE fields.
+
+
 2.5.2
 =====